requirements.txt
APScheduler
boto3
elasticsearch
Flask-RESTful
flask-swagger
marshmallow
PyYAML
setuptools-scm
uwsgidecorators
uwsgitop
